refactor(web): 훅 분리
